Add bsp makefile to edge2
This bsp makefile is used to generate the ```libam_bsp.a``` archive. That archive contains all the bsp code implementations. The bsp header ```am_bsp.h``` still needs to be included manually, but not all the individual sources for the bsp.

The ```?=``` assignments allow this makefile to be called directly or via another makefile that defines the path to the SDK and the Board